# health-tracker-docs/yuanqi/health-tracker-openapi.yaml
openapi: 3.0.0
info:
  title: 健康管理小程序 API
  version: 1.0.0
  description: 供腾讯元器插件导入的 OpenAPI 文档（简化版）
servers:
  - url: "http://139.196.155.234:8080"
    description: "生产环境 API 服务器"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: ok
        data:
          type: object
...
    GoalAddRequest:
      type: object
      required: [userId, goalType, targetValue, period]
      properties:
        userId:
          type: integer
        goalType:
          type: integer
          description: "1=步数,2=睡眠,3=饮食热量"
        targetValue:
          type: number
        currentValue:
          type: number
        period:
          type: string
          example: day
...
    FamilyAddRequest:
      type: object
      required: [userId, name, relation]
      properties:
        userId:
          type: integer
        name:
          type: string
        relation:
          type: string
        age:
          type: integer
        conditionText:
          type: string
...
security:
  - bearerAuth: []
paths:
  /api/auth/mini/login:
    post:
      summary: 微信登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
      responses:
        "200":
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /api/auth/mini/werun:
    post:
      summary: 微信步数同步
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, encryptedData, iv]
              properties:
                code:
                  type: string
                encryptedData:
                  type: string
                iv:
                  type: string
      responses:
        "200":
          description: 同步成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
...
  /api/goal/list:
    get:
      summary: 目标列表
      parameters:
        - in: query
          name: userId
          required: true
          schema:
            type: integer
        - in: query
          name: period
          required: true
          schema:
            type: string
            example: day
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /api/goal/add:
    post:
      summary: 新增目标
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalAddRequest"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /api/goal/update:
    post:
      summary: 更新目标
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/GoalAddRequest"
                - type: object
                  required: [id]
                  properties:
                    id:
                      type: integer
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
...
  /api/family/list:
    get:
      summary: 家庭成员列表
      parameters:
        - in: query
          name: userId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /api/family/add:
    post:
      summary: 新增家庭成员
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FamilyAddRequest"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /api/family/update:
    post:
      summary: 更新家庭成员
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/FamilyAddRequest"
                - type: object
                  required: [id]
                  properties:
                    id:
                      type: integer
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
...